LPS-128599 Move the implementation as inner classes of JDKLoggerTestUtil
